@@ -2,7 +2,7 @@ module TimelineAction
   extend ActiveSupport::Concern
 
   included do
-    before_filter :load_timeline_post, only: :timeline
+    before_action :load_timeline_post, only: :timeline
   end
 
   def timeline
