@@ -1,6 +1,7 @@
 class UsersController < ApplicationController
-  before_filter :load_and_verify_slug, only: [:show, :answers, :questions]
-  before_filter :require_login, except: [:index, :show, :answers, :questions, :new, :create]
+  skip_before_action :check_for_orphaned_authorization, only: [:create]
+  before_action :load_and_verify_slug, only: [:show, :answers, :questions]
+  before_action :require_login, except: [:index, :show, :answers, :questions, :new, :create]
 
   def index
     @users = User.order('reputation DESC').page(params[:page]).per(20)
@@ -33,11 +34,11 @@ class UsersController < ApplicationController
     @questions = @user.questions.includes(:tags, :last_active_user).order('vote_count DESC').page(params[:page]).per(15)
   end
 
-  def new
-  end
-
   def create
-    @user = User.create(user_params)
+    @user = User.new(user_params)
+    @user.authorizations << Authorization.find(session[:auth_id])
+    @user.save! # TODO: handle validation errors - email?
+    session[:auth_id] = nil
     session[:user_id] = @user.id
     redirect_to "/"
   end
