@@ -2,36 +2,24 @@ class User < ActiveRecord::Base
   include Slugger
 
   has_many :answers
-  has_many :authorizations
+  has_many :authorizations, dependent: :destroy
   has_many :badges
   has_many :comments
   has_many :questions
   has_many :reputation_events
   has_many :votes
 
-  accepts_nested_attributes_for :authorizations, allow_destroy: false, reject_if: proc { |obj| obj.blank? }
-
   is_slugged :name
 
   def self.find_by_hash(auth_hash)
-    if auth = Authorization.where('uid = ?', auth_hash[:uid]).where('provider = ?', auth_hash[:provider]).first
-      auth.user
-    else
-      nil
-    end
+    Authorization.find_by_hash(auth_hash).try(:user)
   end
 
   def self.new_from_hash(auth_hash)
-    user = User.new(
+    new(
       email: auth_hash[:info][:email],
       name: auth_hash[:info][:name]
     )
-    user.authorizations.new(
-      email: auth_hash[:info][:email],
-      provider: auth_hash[:provider],
-      uid: auth_hash[:uid]
-    )
-    user
   end
 
   def answer_count
