@@ -16,7 +16,7 @@ ActiveRecord::Schema.define(version: 20141208134739) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "answers", force: true do |t|
+  create_table "answers", force: :cascade do |t|
     t.integer  "question_id"
     t.integer  "user_id"
     t.text     "body"
@@ -27,16 +27,16 @@ ActiveRecord::Schema.define(version: 20141208134739) do
 
   add_index "answers", ["question_id"], name: "index_answers_on_question_id", using: :btree
 
-  create_table "authorizations", force: true do |t|
+  create_table "authorizations", force: :cascade do |t|
     t.integer  "user_id"
-    t.string   "provider"
+    t.string   "provider",   limit: 255
     t.text     "uid"
-    t.string   "email"
+    t.string   "email",      limit: 255
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "badges", force: true do |t|
+  create_table "badges", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "subject_id"
     t.integer  "subject_type"
@@ -47,11 +47,11 @@ ActiveRecord::Schema.define(version: 20141208134739) do
 
   add_index "badges", ["user_id"], name: "index_badges_on_user_id", using: :btree
 
-  create_table "comments", force: true do |t|
+  create_table "comments", force: :cascade do |t|
     t.text     "body"
     t.integer  "user_id"
     t.integer  "post_id"
-    t.string   "post_type"
+    t.string   "post_type",  limit: 255
     t.datetime "created_at"
     t.datetime "updated_at"
   end
@@ -59,23 +59,23 @@ ActiveRecord::Schema.define(version: 20141208134739) do
   add_index "comments", ["post_id", "post_type"], name: "index_comments_on_post_id_and_post_type", using: :btree
   add_index "comments", ["user_id"], name: "index_comments_on_user_id", using: :btree
 
-  create_table "questions", force: true do |t|
+  create_table "questions", force: :cascade do |t|
     t.integer  "user_id"
-    t.string   "title"
+    t.string   "title",               limit: 255
     t.text     "body"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "vote_count",          default: 0, null: false
+    t.integer  "vote_count",                      default: 0, null: false
     t.integer  "last_active_user_id"
     t.datetime "last_active_at"
-    t.integer  "answers_count",       default: 0, null: false
+    t.integer  "answers_count",                   default: 0, null: false
     t.integer  "accepted_answer_id"
   end
 
-  create_table "reputation_events", force: true do |t|
+  create_table "reputation_events", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "event_type"
-    t.string   "action_type"
+    t.string   "action_type", limit: 255
     t.integer  "action_id"
     t.datetime "created_at"
     t.datetime "updated_at"
@@ -83,7 +83,7 @@ ActiveRecord::Schema.define(version: 20141208134739) do
 
   add_index "reputation_events", ["user_id"], name: "index_reputation_events_on_user_id", using: :btree
 
-  create_table "taggings", force: true do |t|
+  create_table "taggings", force: :cascade do |t|
     t.integer  "tag_id"
     t.integer  "question_id"
     t.datetime "created_at"
@@ -93,13 +93,13 @@ ActiveRecord::Schema.define(version: 20141208134739) do
   add_index "taggings", ["question_id"], name: "index_taggings_on_question_id", using: :btree
   add_index "taggings", ["tag_id"], name: "index_taggings_on_tag_id", using: :btree
 
-  create_table "tags", force: true do |t|
-    t.string   "name"
+  create_table "tags", force: :cascade do |t|
+    t.string   "name",       limit: 255
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "timeline_actors", force: true do |t|
+  create_table "timeline_actors", force: :cascade do |t|
     t.integer  "timeline_event_id"
     t.integer  "user_id"
     t.datetime "created_at"
@@ -109,37 +109,37 @@ ActiveRecord::Schema.define(version: 20141208134739) do
   add_index "timeline_actors", ["timeline_event_id"], name: "index_timeline_actors_on_timeline_event_id", using: :btree
   add_index "timeline_actors", ["user_id"], name: "index_timeline_actors_on_user_id", using: :btree
 
-  create_table "timeline_events", force: true do |t|
+  create_table "timeline_events", force: :cascade do |t|
     t.integer  "post_id"
-    t.string   "post_type"
+    t.string   "post_type",  limit: 255
     t.integer  "action"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "users", force: true do |t|
-    t.string   "name"
-    t.string   "email"
+  create_table "users", force: :cascade do |t|
+    t.string   "name",         limit: 255
+    t.string   "email",        limit: 255
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "admin",        default: false, null: false
-    t.integer  "reputation",   default: 0,     null: false
-    t.text     "about_me",     default: "",    null: false
-    t.boolean  "moderator",    default: false, null: false
-    t.integer  "bronze_count", default: 0,     null: false
-    t.integer  "silver_count", default: 0,     null: false
-    t.integer  "gold_count",   default: 0,     null: false
-    t.string   "location",     default: "",    null: false
-    t.string   "website",      default: "",    null: false
+    t.boolean  "admin",                    default: false, null: false
+    t.integer  "reputation",               default: 0,     null: false
+    t.text     "about_me",                 default: "",    null: false
+    t.boolean  "moderator",                default: false, null: false
+    t.integer  "bronze_count",             default: 0,     null: false
+    t.integer  "silver_count",             default: 0,     null: false
+    t.integer  "gold_count",               default: 0,     null: false
+    t.string   "location",                 default: "",    null: false
+    t.string   "website",                  default: "",    null: false
   end
 
-  create_table "votes", force: true do |t|
+  create_table "votes", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "post_id"
     t.integer  "vote_type"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "post_type"
+    t.string   "post_type",  limit: 255
   end
 
   add_index "votes", ["post_type", "post_id"], name: "index_votes_on_post_type_and_post_id", using: :btree
