@@ -2,6 +2,7 @@ class Spinach::Features::EditQuestion < Spinach::FeatureSteps
   include SharedAuthentication
   include SharedPaths
   include SharedQuestion
+  include SharedTagInterface
 
   step 'I am logged in and cannot edit questions' do
     QuestionPolicy.any_instance.stubs(:edit?).returns(false)
@@ -14,11 +15,11 @@ class Spinach::Features::EditQuestion < Spinach::FeatureSteps
   end
 
   step 'I cannot see a link to edit the question' do
-    should_not have_link('edit', href: edit_question_path(current_question.id))
+    refute has_link?('edit', href: edit_question_path(current_question.id))
   end
 
   step 'I can see a link to edit the question' do
-    should have_link('edit', href: edit_question_path(current_question.id))
+    assert has_link?('edit', href: edit_question_path(current_question.id))
   end
 
   step 'I visit the question edit page' do
@@ -28,16 +29,16 @@ class Spinach::Features::EditQuestion < Spinach::FeatureSteps
   step 'I edit and submit the question data' do
     fill_in 'question_body', with: 'this is an edited question body'
     fill_in 'question_title', with: 'new question title'
-    fill_in 'question_tag_list', with: 'new-list, tags'
+    input_and_add_tags('new-list, tags')
     find(:xpath, '//input[@name="commit"]').click
   end
 
   step 'I am on the question page' do
-    current_path.should == question_path(Question.first)
+    assert_path question_path(Question.first)
   end
 
   step 'I should see the updated question' do
-    should have_content 'this is an edited question body'
-    should have_content 'new question title'
+    assert_text 'this is an edited question body'
+    assert_text 'new question title'
   end
 end
