@@ -6,7 +6,6 @@ end
 ENV['RAILS_ENV'] = 'test'
 require './config/environment'
 
-require 'rspec/expectations'
 require 'mocha/setup'
 
 # Require all shared step files
@@ -14,10 +13,7 @@ Dir["#{Rails.root}/features/steps/shared/*.rb"].each {|file| require file}
 
 # Setup Javascript tags
 require 'capybara/poltergeist'
-Capybara.javascript_driver = :poltergeist
-Spinach.hooks.on_tag("javascript") do
-  ::Capybara.current_driver = ::Capybara.javascript_driver
-end
+Capybara.default_driver = :poltergeist
 Capybara.default_wait_time = 10
 
 # Setup Omniauth mock
@@ -32,16 +28,28 @@ def omniauth_hash
   }
 end
 OmniAuth.config.test_mode = true
-OmniAuth.config.add_mock(:google, omniauth_hash)
+OmniAuth.config.mock_auth[:google] = omniauth_hash
 
 # DBCleaner
 require 'database_cleaner'
 DatabaseCleaner.strategy = :truncation
 
 Spinach.hooks.before_scenario { DatabaseCleaner.clean }
-Spinach.hooks.after_scenario { Mocha::Mockery.teardown }
+Spinach.hooks.after_scenario do
+  Mocha::Mockery.teardown
+  OmniAuth.config.mock_auth[:google] = omniauth_hash # reset any admin trickery
+end
 
 # factorygirl niceness
 Spinach.hooks.before_run do
   include FactoryGirl::Syntax::Methods
+end
+
+class Spinach::FeatureSteps
+  attr_accessor :assertions
+
+  def initialize
+    super
+    @assertions = 0
+  end
 end
