@@ -50,7 +50,6 @@ module QA
         answers = posts.select { |p| p["PostTypeId"] == "2" }
 
         grouped_edits = build_edits(post_histories)
-        post_histories = nil
         posts = nil
 
         # create posts
@@ -73,9 +72,7 @@ module QA
         voterow = Nokogiri::XML::Document.parse(File.read("#{@dir}/votes.xml")).css('votes row')
         puts " Creating votes"
 
-        size = User.count
         user_ids = User.pluck(:id)
-        votes = []
         bar = progress_bar('Votes', voterow.count)
         @conn.exec('COPY votes (user_id, post_id, post_type, vote_type, created_at, updated_at) FROM STDIN WITH CSV')
 
@@ -139,11 +136,11 @@ module QA
         print " Updating cache counters & creating background jobs for badges"
         Question.all.each do |q|
           Jobs::QuestionStats.new.perform(q.id)
-          Jobs::Badge.perform_async(:question_vote, q.global_id)
+          Jobs::Badge.perform_async(:question_vote, q.to_global_id)
         end
         Answer.all.each do |a|
           Jobs::AnswerStats.new.perform(a.id)
-          Jobs::Badge.perform_async(:answer_vote, a.global_id)
+          Jobs::Badge.perform_async(:answer_vote, a.to_global_id)
         end
         puts " - done!"
       end
@@ -163,7 +160,7 @@ module QA
         puts " Grouping GUIDs into single edits"
         bar = progress_bar('Grouping', guidgroups.length)
         groupededits = Hash.new { |hash, key| hash[key] = [] }
-        guidgroups.each do |key, edit|
+        guidgroups.each do |_, edit|
           eobj = StackExchange::Edit.new(edit)
           bar.increment
           groupededits[eobj.post_id] << eobj
@@ -180,7 +177,6 @@ module QA
         bar = progress_bar('Questions', questions.length)
         questions.each do |q|
           bar.increment
-          qu = Question.new
           edits = grouped_edits[q['Id']]
           originator = edits.select(&:new_record)[0]
           edits.delete originator
