@@ -1,14 +1,19 @@
 require 'spec_helper'
 
-describe AdminController do
+describe Admin::DashboardController do
   context 'while logged in as an administrator' do
     before(:each) { sign_in(admin) }
 
     context '#index' do
       it 'returns success' do
         get :index
-        expect(response.status).to eq(200)
+        expect(response).to be_success
       end
+    end
+
+    it 'health returns success' do
+      get :health
+      expect(response).to be_success
     end
   end
 end
